requirements.txt
flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate

python-dotenv
python-json-logger

pytest
black
flake8
flake8-blind-except
flake8-isort
isort
pep8-naming